# ⚡️ Speed up function `outputs_to_objects` by 88%