--- a/unstructured_inference/__version__.py
+++ b/unstructured_inference/__version__.py
@@ -1,1 +1,1 @@
-__version__ = "1.0.7"  # pragma: no cover
+__version__ = "1.0.8-dev0"  # pragma: no cover